refactor: move unit test actions to dedicated route files
