+++ packages/insomnia/src/ui/routes/actions.tsx
@@ -38,119 +38,6 @@ export function safeToUseInsomniaFileNameWithExt(fileName: string) {
   return `${safeToUseInsomniaFileName(fileName)}.yaml`;
 }
 
-// Unit Test
-
-export const deleteTestAction: ActionFunction = async ({ params }) => {
-  const { testId } = params;
-  invariant(typeof testId === 'string', 'Test ID is required');
-
-  const unitTest = await database.getWhere<UnitTest>(models.unitTest.type, {
-    _id: testId,
-  });
-
-  invariant(unitTest, 'Test not found');
-
-  await models.unitTest.remove(unitTest);
-  window.main.trackSegmentEvent({ event: SegmentEvent.unitTestDelete });
-
-  return null;
-};
-
-export const updateTestAction: ActionFunction = async ({ request, params }) => {
-  const { testId } = params;
-  const data = (await request.json()) as Partial<UnitTest>;
-
-  const unitTest = await database.getWhere<UnitTest>(models.unitTest.type, {
-    _id: testId,
-  });
-  invariant(unitTest, 'Test not found');
-
-  await models.unitTest.update(unitTest, data);
-
-  return null;
-};
-
-export const runTestAction: ActionFunction = async ({ params }) => {
-  const { organizationId, projectId, workspaceId, testSuiteId, testId } = params;
-  invariant(typeof testId === 'string', 'Test ID is required');
-
-  const unitTest = await database.getWhere<UnitTest>(models.unitTest.type, {
-    _id: testId,
-  });
-  invariant(unitTest, 'Test not found');
-
-  const tests: Test[] = [
-    {
-      name: unitTest.name,
-      code: unitTest.code,
-      defaultRequestId: unitTest.requestId,
-    },
-  ];
-  const src = generate([{ name: 'My Suite', suites: [], tests }]);
-
-  const sendRequest = getSendRequestCallback();
-
-  let results: TestResults = {
-    failures: [],
-    passes: [],
-    pending: [],
-    stats: {
-      suites: 0,
-      tests: 0,
-      passes: 0,
-      pending: 0,
-      failures: 0,
-      start: undefined,
-      end: undefined,
-      duration: undefined,
-    },
-    tests: [],
-  };
-
-  try {
-    results = await runTests(src, { sendRequest });
-    const testResult = await models.unitTestResult.create({
-      results,
-      parentId: unitTest.parentId,
-    });
-    window.main.trackSegmentEvent({ event: SegmentEvent.unitTestRun, properties: { organizationId, projectId } });
-
-    return redirect(
-      `/organization/${organizationId}/project/${projectId}/workspace/${workspaceId}/test/test-suite/${testSuiteId}/test-result/${testResult._id}`,
-    );
-  } catch (error) {
-    // create a result manually so that it can be displayed in the UI
-    results.stats.failures = 1;
-    results.stats.tests = 1;
-    results.tests.push({
-      currentRetry: 0,
-      duration: 0,
-      err: {
-        actual: undefined,
-        expected: undefined,
-        message: error.toString(),
-        multiple: [],
-        operator: undefined,
-        showDiff: false,
-        stack: '',
-      },
-      file: '',
-      fullTitle: unitTest.name,
-      id: '',
-      title: unitTest.name,
-    });
-    const testResult = await models.unitTestResult.create({
-      results,
-      parentId: unitTest.parentId,
-    });
-    window.main.trackSegmentEvent({ event: SegmentEvent.unitTestRun, properties: { organizationId, projectId } });
-
-    return redirect(
-      `/organization/${organizationId}/project/${projectId}/workspace/${workspaceId}/test/test-suite/${testSuiteId}/test-result/${testResult._id}`,
-    );
-  }
-};
-
 // Api Spec
 export const updateApiSpecAction: ActionFunction = async ({ request, params }) => {
   const { workspaceId } = params;
